+++ src/faebryk/library/components/Crystal_Oscillator.py
