+++ src/faebryk/library/traits/has_single_electric_reference.py
@@ -8,7 +8,7 @@ from typing import TYPE_CHECKING
 from faebryk.core.trait import Trait
 
 if TYPE_CHECKING:
-    from faebryk.library.ElectricPower import ElectricPower
+    from faebryk.library.interfaces.ElectricPower import ElectricPower
 
 
 class has_single_electric_reference(Trait):
